Fixed a copy paste bug and other things. However some of the examples, especially mesh 3d have failed through the time.

+++ src/common/mathutil.c
@@ -17,11 +17,11 @@ int isqrt(int x) {
     while (q > 1) {
         int t;
         q >>= 2;
-        t = x - r - q;
+        t = (int)(x - r - q);
         r >>= 1;
         if (t >= 0) {
             x = t;
-            r += q;
+            r += (int)q;
         }
     }
     return r;
